Fixed script pathing, functioning on distributed training
