--- dna-diffusion/README.md
@@ -1,7 +1,0 @@
-# DNA-diffusion files structure
-
-- data: contains the data used in the DNA diffusion project
-- losses: contains the losses used in the DNA diffusion project
-- metrics: contains the internal metrics used to evaluate the quality of the generated sequences after training
-- models: contains the models used in the DNA diffusion project (UNET, VQ-VAE, etc.)
-- utils: contains the utils used in the DNA diffusion project
