--- dna-diffusion/metrics/README.md
@@ -1,1 +1,0 @@
-# Internal metrics to assess the quality of generated sequences after training.
