--- dna-diffusion/models/README.md
@@ -1,1 +1,0 @@
-# Models
