--- dna-diffusion/requirements.yml
@@ -1,22 +1,0 @@
-name: dnadiffusion
-channels:
-  - nodefaults
-  - conda-forge
-  - bioconda
-dependencies:
-  - pytorch-gpu
-  - torchvision
-  - accelerate
-  - jupyterlab
-  - matplotlib
-  - pandas
-  - seaborn
-  - einops
-  - wandb
-  - genomepy
-  - pytorch-lightning
-  - hydra-core
-  - pip
-  - pip:
-      - gimmemotifs
-      - memory-efficient-attention-pytorch
