--- dna-diffusion/utils/scheduler.py
@@ -1,32 +1,0 @@
-import torch
-
-
-def linear_beta_schedule(timesteps: int, beta_end: float = 0.005):
-    beta_start = 0.0001
-
-    return torch.linspace(beta_start, beta_end, timesteps)
-
-
-def cosine_beta_schedule(timesteps: int, s: float = 0.008):
-    """
-    cosine schedule as proposed in https://arxiv.org/abs/2102.09672
-    """
-    steps = timesteps + 1
-    x = torch.linspace(0, timesteps, steps)
-    alphas_cumprod = torch.cos(((x / timesteps) + s) / (1 + s) * torch.pi * 0.5) ** 2
-    alphas_cumprod = alphas_cumprod / alphas_cumprod[0]
-    betas = 1 - (alphas_cumprod[1:] / alphas_cumprod[:-1])
-    return torch.clip(betas, 0.0001, 0.9999)
-
-
-def quadratic_beta_schedule(timesteps: int):
-    beta_start = 0.0001
-    beta_end = 0.02
-    return torch.linspace(beta_start**0.5, beta_end**0.5, timesteps) ** 2
-
-
-def sigmoid_beta_schedule(timesteps: int):
-    beta_start = 0.001
-    beta_end = 0.02
-    betas = torch.linspace(-6, 6, timesteps)
-    return torch.sigmoid(betas) * (beta_end - beta_start) + beta_start
