+++ dnadiffusion/models/networks.py
@@ -1,0 +1,465 @@
+import math
+from functools import partial
+from typing import Optional
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from einops import rearrange
+from memory_efficient_attention_pytorch import Attention as EfficientAttention
+from torch import einsum
+
+from dnadiffusion.utils.utils import default, exists, l2norm
+
+
+class SinusoidalPositionEmbeddings(nn.Module):
+    def __init__(self, dim: int) -> None:
+        super().__init__()
+        self.dim = dim
+
+    def forward(self, time: torch.Tensor):
+        half_dim = self.dim // 2
+        embeddings = math.log(10000) / (half_dim - 1)
+        embeddings = torch.exp(torch.arange(half_dim) * -embeddings)
+
+        embeddings = time[:, None] * embeddings[None, :]
+        embeddings = torch.cat((embeddings.sin(), embeddings.cos()), dim=-1)
+        return embeddings
+
+
+class ResBlock(nn.Module):
+
+    """
+    Iniialize a residual block with two convolutions followed by batchnorm layers
+    """
+
+    def __init__(self, in_size: int, hidden_size: int, out_size: int) -> None:
+        super().__init__()
+        self.conv1 = nn.Conv2d(in_size, hidden_size, 3, padding=1)
+        self.conv2 = nn.Conv2d(hidden_size, out_size, 3, padding=1)
+        self.batchnorm1 = nn.BatchNorm2d(hidden_size)
+        self.batchnorm2 = nn.BatchNorm2d(out_size)
+
+    def convblock(self, x: torch.Tensor):
+        x = F.relu(self.batchnorm1(self.conv1(x)))
+        x = F.relu(self.batchnorm2(self.conv2(x)))
+        return x
+
+    """
+    Combine output with the original input
+    """
+
+    def forward(self, x):
+        return x + self.convblock(x)
+
+
+class ConvBlock_2d(nn.Module):
+    def __init__(self, in_channels: int, out_channels: int) -> None:
+        super().__init__()
+
+        self.conv1 = nn.Sequential(
+            nn.Conv2d(in_channels, out_channels, 4, padding=2),
+            nn.BatchNorm2d(out_channels),
+            nn.ReLU(),
+        )
+        self.conv2 = nn.Sequential(
+            nn.Conv2d(out_channels, out_channels, 4, 1, 1),
+            nn.BatchNorm2d(out_channels),
+            nn.ReLU(),
+        )
+
+        self._init_weights()
+
+    def _init_weights(self):
+        for m in self.modules():
+            if isinstance(m, nn.Conv2d):
+                nn.init.kaiming_normal_(m.weight)
+                if m.bias is not None:
+                    nn.init.zeros_(m.bias)
+            elif isinstance(m, nn.BatchNorm2d):
+                nn.init.constant_(m.weight, 1)
+                nn.init.zeros_(m.bias)
+
+    def forward(self, x: torch.Tensor):
+        x = self.conv1(x)
+        x = self.conv2(x)
+
+        return x
+
+
+class EmbedFC(nn.Module):
+    def __init__(self, input_dim: int, emb_dim: int) -> None:
+        super().__init__()
+        """
+        generic one layer FC NN for embedding things
+        """
+        self.input_dim = input_dim
+        layers = [nn.Linear(input_dim, emb_dim), nn.GELU(), nn.Linear(emb_dim, emb_dim)]
+        self.model = nn.Sequential(*layers)
+
+    def forward(self, x: torch.Tensor):
+        return self.model(x)
+
+
+class Residual(nn.Module):
+    def __init__(self, fn) -> None:
+        super().__init__()
+        self.fn = fn
+
+    def forward(self, x: torch.Tensor, *args, **kwargs):
+        return self.fn(x, *args, **kwargs) + x
+
+
+def Upsample(dim: int, dim_out: Optional[int]):
+    return nn.Sequential(
+        nn.Upsample(scale_factor=2, mode="nearest"),
+        nn.Conv2d(dim, default(dim_out, dim), 3, padding=1),
+    )
+
+
+def Downsample(dim: int, dim_out: Optional[int]):
+    return nn.Conv2d(dim, default(dim_out, dim), 4, 2, 1)
+
+
+class LayerNorm(nn.Module):
+    def __init__(self, dim: int) -> None:
+        super().__init__()
+        self.g = nn.Parameter(torch.ones(1, dim, 1, 1))
+
+    def forward(self, x):
+        eps = 1e-5 if x.dtype == torch.float32 else 1e-3
+        var = torch.var(x, dim=1, unbiased=False, keepdim=True)
+        mean = torch.mean(x, dim=1, keepdim=True)
+        return (x - mean) * (var + eps).rsqrt() * self.g
+
+
+class PreNorm(nn.Module):
+    def __init__(self, dim: int, fn: nn.Module) -> None:
+        super().__init__()
+        self.fn = fn
+        self.norm = LayerNorm(dim)
+
+    def forward(self, x):
+        x = self.norm(x)
+        return self.fn(x)
+
+
+# positional embeds
+
+
+class LearnedSinusoidalPosEmb(nn.Module):
+    """following @crowsonkb 's lead with learned sinusoidal pos emb"""
+
+    """ https://github.com/crowsonkb/v-diffusion-jax/blob/master/diffusion/models/danbooru_128.py#L8 """
+
+    def __init__(self, dim: int) -> None:
+        super().__init__()
+        assert (dim % 2) == 0
+        half_dim = dim // 2
+        self.weights = nn.Parameter(torch.randn(half_dim))
+
+    def forward(self, x: torch.Tensor):
+        x = rearrange(x, "b -> b 1")
+        freqs = x * rearrange(self.weights, "d -> 1 d") * 2 * math.pi
+        fouriered = torch.cat((freqs.sin(), freqs.cos()), dim=-1)
+        fouriered = torch.cat((x, fouriered), dim=-1)
+        return fouriered
+
+
+# building block modules
+
+
+class Block(nn.Module):
+    def __init__(self, dim: int, dim_out: int, groups: int = 8) -> None:
+        super().__init__()
+        self.proj = nn.Conv2d(dim, dim_out, 3, padding=1)
+        self.norm = nn.GroupNorm(groups, dim_out)
+        self.act = nn.SiLU()
+
+    def forward(self, x: torch.Tensor, scale_shift: Optional[torch.Tensor] = None):
+        x = self.proj(x)
+        x = self.norm(x)
+
+        if exists(scale_shift):
+            scale, shift = scale_shift
+            x = x * (scale + 1) + shift
+
+        x = self.act(x)
+        return x
+
+
+class ResnetBlock(nn.Module):
+    def __init__(
+        self, dim: int, dim_out: int, *, time_emb_dim: Optional[int], groups: int = 8
+    ) -> None:
+        super().__init__()
+        self.mlp = (
+            (nn.Sequential(nn.SiLU(), nn.Linear(time_emb_dim, dim_out * 2)))
+            if exists(time_emb_dim)
+            else None
+        )
+
+        self.block1 = Block(dim, dim_out, groups=groups)
+        self.block2 = Block(dim_out, dim_out, groups=groups)
+        self.res_conv = nn.Conv2d(dim, dim_out, 1) if dim != dim_out else nn.Identity()
+
+    def forward(self, x: torch.Tensor, time_emb: Optional[torch.Tensor] = None):
+        scale_shift = None
+        if exists(self.mlp) and exists(time_emb):
+            time_emb = self.mlp(time_emb)
+            time_emb = rearrange(time_emb, "b c -> b c 1 1")
+            scale_shift = time_emb.chunk(2, dim=1)
+
+        h = self.block1(x, scale_shift=scale_shift)
+
+        h = self.block2(h)
+
+        return h + self.res_conv(x)
+
+
+class LinearAttention(nn.Module):
+    def __init__(self, dim: int, heads: int = 4, dim_head: int = 32) -> None:
+        super().__init__()
+        self.scale = dim_head**-0.5
+        self.heads = heads
+        hidden_dim = dim_head * heads
+        self.to_qkv = nn.Conv2d(dim, hidden_dim * 3, 1, bias=False)
+        self.to_out = nn.Sequential(nn.Conv2d(hidden_dim, dim, 1), LayerNorm(dim))
+
+    def forward(self, x: torch.Tensor):
+        b, c, h, w = x.shape
+        qkv = self.to_qkv(x).chunk(3, dim=1)
+        q, k, v = map(
+            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv
+        )
+
+        q = q.softmax(dim=-2)
+        k = k.softmax(dim=-1)
+
+        q = q * self.scale
+        v = v / (h * w)
+
+        context = torch.einsum("b h d n, b h e n -> b h d e", k, v)
+
+        out = torch.einsum("b h d e, b h d n -> b h e n", context, q)
+        out = rearrange(out, "b h c (x y) -> b (h c) x y", h=self.heads, x=h, y=w)
+        return self.to_out(out)
+
+
+class Attention(nn.Module):
+    def __init__(
+        self, dim: int, heads: int = 4, dim_head: int = 32, scale: int = 10
+    ) -> None:
+        super().__init__()
+        self.scale = scale
+        self.heads = heads
+        hidden_dim = dim_head * heads
+        self.to_qkv = nn.Conv2d(dim, hidden_dim * 3, 1, bias=False)
+        self.to_out = nn.Conv2d(hidden_dim, dim, 1)
+
+    def forward(self, x: torch.Tensor):
+        b, c, h, w = x.shape
+        qkv = self.to_qkv(x).chunk(3, dim=1)
+        q, k, v = map(
+            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv
+        )
+
+        q, k = map(l2norm, (q, k))
+
+        sim = einsum("b h d i, b h d j -> b h i j", q, k) * self.scale
+        attn = sim.softmax(dim=-1)
+        out = einsum("b h i j, b h d j -> b h i d", attn, v)
+        out = rearrange(out, "b h (x y) d -> b (h d) x y", x=h, y=w)
+        return self.to_out(out)
+
+
+class CrossAttention_lucas(nn.Module):
+    def __init__(
+        self, dim: int, heads: int = 1, dim_head: int = 32, scale: int = 10
+    ) -> None:
+        super().__init__()
+        self.scale = scale
+        self.heads = heads
+        hidden_dim = dim_head * heads
+        self.to_qkv = nn.Conv2d(dim, hidden_dim * 3, 1, bias=False)
+        self.to_out = nn.Conv2d(hidden_dim, dim, 1)
+
+    def forward(self, x: torch.Tensor, y: torch.Tensor):
+        b, c, h, w = x.shape
+        b_y, c_y, h_y, w_y = y.shape
+
+        qkv_x = self.to_qkv(x).chunk(3, dim=1)
+        qkv_y = self.to_qkv(y).chunk(3, dim=1)
+
+        q_x, k_x, v_x = map(
+            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv_x
+        )
+
+        q_y, k_y, v_y = map(
+            lambda t: rearrange(t, "b (h c) x y -> b h c (x y)", h=self.heads), qkv_y
+        )
+
+        q, k = map(l2norm, (q_x, k_y))
+
+        sim = einsum("b h d i, b h d j -> b h i j", q, k) * self.scale
+        attn = sim.softmax(dim=-1)
+        out = einsum("b h i j, b h d j -> b h i d", attn, v_y)
+        out = rearrange(out, "b h (x y) d -> b (h d) x y", x=h, y=w)
+        return self.to_out(out)
+
+
+class Unet_lucas(nn.Module):
+    def __init__(
+        self,
+        dim: int,
+        init_dim: Optional[int] = None,
+        dim_mults: tuple = (1, 2, 4),
+        channels: int = 1,
+        resnet_block_groups: int = 8,
+        learned_sinusoidal_dim: int = 18,
+        num_classes: int = 10,
+        output_attention: bool = False,
+    ) -> None:
+        super().__init__()
+
+        # determine dimensions
+
+        channels = 1
+        self.channels = channels
+        # if you want to do self conditioning uncomment this
+        input_channels = channels
+        self.output_attention = output_attention
+
+        init_dim = default(init_dim, dim)
+        self.init_conv = nn.Conv2d(input_channels, init_dim, (7, 7), padding=3)
+        dims = [init_dim, *map(lambda m: dim * m, dim_mults)]
+
+        in_out = list(zip(dims[:-1], dims[1:]))
+        block_klass = partial(ResnetBlock, groups=resnet_block_groups)
+
+        # time embeddings
+        time_dim = dim * 4
+
+        sinu_pos_emb = LearnedSinusoidalPosEmb(learned_sinusoidal_dim)
+        fourier_dim = learned_sinusoidal_dim + 1
+
+        self.time_mlp = nn.Sequential(
+            sinu_pos_emb,
+            nn.Linear(fourier_dim, time_dim),
+            nn.GELU(),
+            nn.Linear(time_dim, time_dim),
+        )
+
+        if num_classes is not None:
+            self.label_emb = nn.Embedding(num_classes, time_dim)
+
+        # layers
+        self.downs = nn.ModuleList([])
+        self.ups = nn.ModuleList([])
+        num_resolutions = len(in_out)
+
+        for ind, (dim_in, dim_out) in enumerate(in_out):
+            is_last = ind >= (num_resolutions - 1)
+
+            self.downs.append(
+                nn.ModuleList(
+                    [
+                        block_klass(dim_in, dim_in, time_emb_dim=time_dim),
+                        block_klass(dim_in, dim_in, time_emb_dim=time_dim),
+                        Residual(PreNorm(dim_in, LinearAttention(dim_in))),
+                        Downsample(dim_in, dim_out)
+                        if not is_last
+                        else nn.Conv2d(dim_in, dim_out, 3, padding=1),
+                    ]
+                )
+            )
+
+        mid_dim = dims[-1]
+        self.mid_block1 = block_klass(mid_dim, mid_dim, time_emb_dim=time_dim)
+        self.mid_attn = Residual(PreNorm(mid_dim, Attention(mid_dim)))
+        self.mid_block2 = block_klass(mid_dim, mid_dim, time_emb_dim=time_dim)
+
+        for ind, (dim_in, dim_out) in enumerate(reversed(in_out)):
+            is_last = ind == (len(in_out) - 1)
+            self.ups.append(
+                nn.ModuleList(
+                    [
+                        block_klass(dim_out + dim_in, dim_out, time_emb_dim=time_dim),
+                        block_klass(dim_out + dim_in, dim_out, time_emb_dim=time_dim),
+                        Residual(PreNorm(dim_out, LinearAttention(dim_out))),
+                        Upsample(dim_out, dim_in)
+                        if not is_last
+                        else nn.Conv2d(dim_out, dim_in, 3, padding=1),
+                    ]
+                )
+            )
+
+        self.final_res_block = block_klass(dim * 2, dim, time_emb_dim=time_dim)
+        self.final_conv = nn.Conv2d(dim, 1, 1)
+        self.cross_attn = EfficientAttention(
+            dim=200,
+            dim_head=64,
+            heads=1,
+            memory_efficient=True,
+            q_bucket_size=1024,
+            k_bucket_size=2048,
+        )
+        self.norm_to_cross = nn.LayerNorm(dim * 4)
+
+    def forward(self, x: torch.Tensor, time: torch.Tensor, classes: torch.Tensor):
+        x = self.init_conv(x)
+        r = x.clone()
+
+        t_start = self.time_mlp(time)
+        t_mid = t_start.clone()
+        t_end = t_start.clone()
+        t_cross = t_start.clone()
+
+        if classes is not None:
+            t_start += self.label_emb(classes)
+            t_mid += self.label_emb(classes)
+            t_end += self.label_emb(classes)
+            t_cross += self.label_emb(classes)
+
+        h = []
+
+        for block1, block2, attn, downsample in self.downs:
+            x = block1(x, t_start)
+            h.append(x)
+
+            x = block2(x, t_start)
+            x = attn(x)
+            h.append(x)
+
+            x = downsample(x)
+
+        x = self.mid_block1(x, t_mid)
+        x = self.mid_attn(x)
+        x = self.mid_block2(x, t_mid)
+
+        for block1, block2, attn, upsample in self.ups:
+            x = torch.cat((x, h.pop()), dim=1)
+            x = block1(x, t_mid)
+
+            x = torch.cat((x, h.pop()), dim=1)
+            x = block2(x, t_mid)
+            x = attn(x)
+
+            x = upsample(x)
+
+        x = torch.cat((x, r), dim=1)
+
+        x = self.final_res_block(x, t_end)
+        x = self.final_conv(x)
+        x_reshaped = x.reshape(-1, 4, 200)
+        t_cross_reshaped = t_cross.reshape(-1, 4, 200)
+
+        crossattention_out = self.cross_attn(
+            self.norm_to_cross(x_reshaped.reshape(-1, 800)).reshape(-1, 4, 200),
+            context=t_cross_reshaped,
+        )  # (-1,1, 4, 200)
+        crossattention_out = x.view(-1, 1, 4, 200)
+        x = x + crossattention_out
+        if self.output_attention:
+            return x, crossattention_out
+        return x
