+++ dnadiffusion/utils/utils.py
@@ -1,0 +1,74 @@
+import math
+
+import numpy as np
+import torch.nn.functional as F
+
+
+def exists(x):
+    return x is not None
+
+
+def cycle(dl):
+    while True:
+        yield from dl
+
+
+def has_int_squareroot(num):
+    return (math.sqrt(num) ** 2) == num
+
+
+def num_to_groups(num, divisor):
+    groups = num // divisor
+    remainder = num % divisor
+    arr = [divisor] * groups
+    if remainder > 0:
+        arr.append(remainder)
+    return arr
+
+
+def convert_image_to(img_type, image):
+    if image.mode != img_type:
+        return image.convert(img_type)
+    return image
+
+
+def l2norm(t):
+    return F.normalize(t, dim=-1)
+
+
+def default(val, d):
+    if exists(val):
+        return val
+    return d() if callable(d) else d
+
+
+def extract(a, t, x_shape, device=None):
+    batch_size = t.shape[0]
+    if device:
+        a = a.to(device)
+        t = t.to(device)
+
+    out = a.gather(-1, t)
+    result = out.reshape(batch_size, *((1,) * (len(x_shape) - 1)))
+    if device:
+        result.to(device)
+    return result
+
+
+def one_hot_encode(seq, alphabet, max_seq_len):
+    """One-hot encode a sequence."""
+    seq_len = len(seq)
+    seq_array = np.zeros((max_seq_len, len(alphabet)))
+    for i in range(seq_len):
+        seq_array[i, alphabet.index(seq[i])] = 1
+    return seq_array
+
+
+def encode(seq, alphabet):
+    """Encode a sequence."""
+    seq_len = len(seq)
+    seq_array = np.zeros(len(alphabet))
+    for i in range(seq_len):
+        seq_array[alphabet.index(seq[i])] = 1
+
+    return seq_array
