+++ train.py
@@ -1,0 +1,6 @@
+from dnadiffusion.data.dataloader import LoadingData
+from dnadiffusion.trainer import Trainer
+
+if __name__ == "__main__":
+    trainer = Trainer()
+    trainer.train()
